+++ include/symmetric.h
@@ -10,10 +10,10 @@
 #ifndef SYM_H
 #define SYM_H
 #include <tomcrypt.h>
-int generateRandom(unsigned char *array, int len);
+int generateRandom(unsigned char *array, unsigned long len);
 
-int symmetricEncrypt(unsigned char *key, int keylen, unsigned char *in, unsigned long len, unsigned char *IV, int ivlen);
+int symmetricEncrypt(unsigned char *key, unsigned long keylen, unsigned char *in, unsigned long len, unsigned char *IV, unsigned long ivlen);
 
-int symmetricDecrypt(unsigned char *key, int keylen, unsigned char *buffer, unsigned long len, unsigned char *IV, int ivlen);
+int symmetricDecrypt(unsigned char *key, unsigned long keylen, unsigned char *buffer, unsigned long len, unsigned char *IV, unsigned long ivlen);
 
 #endif
